Spelling check update

+++ Xecure Password Manager/settings.py
@@ -5,11 +5,14 @@
 #credentials and encrypts the records of the user inside a secure MySQL database.
 
 #This module contains the settings controls that give users control over their data.
+import socket
 
 from clear import clear
 from msvcrt import getch
 from dbsetup import delete_record, delete_all_records, fetch_rec_by_id_gen, checkIfNoRecords
 from passwordRecovery import update_master_password, changeMasterEmail
+from input_val import validateRecordName, validateEmail, validatePassword
+import smtplib
 
 #This function is the settings menu that contains the controls of delete a record, delete all records, change email,
 #change password, change security question (Tentative Feature), and Back to dashboard.
@@ -20,7 +23,7 @@ def settings(userId):
             clear()
             print("--------------Settings------------------")
             # Presents user's with all options available
-            print("\nWhat would you like to do?.\n")
+            print("\nPlease select one of the following options\n")
             print("[1] Delete a Record")
             print("[2] Delete All Records")
             print("[3] Change Master Account Email")
@@ -40,6 +43,7 @@ def settings(userId):
 
             if menuSelection == 4:
                 result = update_master_password(userId)
+
                 if result == True:
                     break
 
@@ -48,13 +52,13 @@ def settings(userId):
                 break
 
             if menuSelection > 4 or menuSelection < 0:
-                print("\nPlease enter a number between 0 and 4.")
+                print("\nPlease enter a number between 0 and 4")
                 print("Press any key to try again...")
                 getch()
                 clear()
 
         except Exception:
-            print("\nInvalid Input.")
+            print("\nInvalid Input")
             print("Press any key to try again...")
             getch()
             clear()
@@ -71,14 +75,12 @@ def deleteRecord(userId):
 
             if checkIfNoRecords(userId) == 0:
                 print("\n**No records in database**")
-                print("Press any key to go back...")
+                print("Press any key to go back to Settings...")
                 getch()
                 clear()
                 break
 
             else:
-
-                #todo implement better input val
                 print("\nPlease enter the name of the record you wish to delete:")
                 recordSelected = input("\nRecord Name: ")
 
@@ -86,19 +88,37 @@ def deleteRecord(userId):
                     return '0'
 
                 if recordSelected.isspace() or recordSelected == "":
-                    print("\nRecord name cannot be empty.\nPress any key to continue...")
+                    print("\nRecord name cannot be empty\nPress any key to continue...")
                     getch()
                     clear()
-                    break
+                    continue
 
-                elif recordSelected.isnumeric():
-                    print("\nRecord name cannot be composed of only numbers.\nPress any key to continue...")
+                if recordSelected.isnumeric():
+                    print("\nRecord name cannot be composed of only numbers\nPress any key to continue...")
                     getch()
                     clear()
-                    break
+                    continue
+
+                recordSelected = validateRecordName(recordSelected)
+
+                if recordSelected == '1':
+                    print("\nRecord name must be at least 3 characters long\nPress any key to try again...")
+                    getch()
+                    continue
+
+                if recordSelected == '2':
+                    print("\nRecord name must be less than 35 characters long\nPress any key to try again...")
+                    getch()
+                    continue
+
+                if recordSelected == '3':
+                    print("\nRecord names must contain letters. Numbers and special characters are optional"
+                          "\nSpecial characters allowed: Underscores (_), spaces ( ), and periods (.)"
+                          "\nPress any key to try again...")
+                    getch()
+                    continue
 
 
-                #todo implement function to find a specific record
                 else:
                     test = fetch_rec_by_id_gen(userId, recordSelected, "Record_Name")
                     if recordSelected == test[2]:
@@ -107,7 +127,6 @@ def deleteRecord(userId):
                         clear()
 
                         confirmation = "0"#Ensures user wants to delete this record
-                        #todo implement better input val
                         while confirmation != "1" or confirmation != "2" or confirmation == "0":
                             print("--------------Delete Record------------------")
                             print("\nAre you sure you want to delete the " + recordSelected + " record?")
@@ -119,24 +138,23 @@ def deleteRecord(userId):
                                 ready = 2#Flag is set to exit function
                                 delete_record(recordSelected, userId)
                                 print("\nRecord " + recordSelected + " was deleted!")
-                                #todo implement record deletion
                                 break
 
                             elif confirmation == "2":#Record will not be deleted if 2
                                 ready = 2#Flag set to exit function
-                                print("\nRecord was not deleted.")
+                                print("\nRecord was not deleted")
                                 break
 
                             else:#Invalid input
-                                print("\nInvalid Input!\nPress any key to try again...")
+                                print("\nInvalid Input\nPress any key to try again...")
                                 getch()
                                 clear()
 
                     else:#If record not fund
-                        print("\nRecord not found!\nPress any key to continue...")
+                        print("\nRecord not found!\nPress any key to try again...")
                         getch()
                         clear()
-                        break
+                        continue
 
             if ready == 2:#Function ready to break
                 print("Press any key to go back to Settings...")
@@ -145,7 +163,7 @@ def deleteRecord(userId):
                 break
 
         except Exception:
-            print("\nInvalid Input.")
+            print("\nInvalid Input")
             print("Press any key to try again...")
             getch()
             clear()
@@ -153,62 +171,58 @@ def deleteRecord(userId):
 
 #Function deletes all record after prompting the user to confirm action
 def deleteAllRecords(userId):
-    clear()
-    confirmation = "0"
-    #todo implement input val
-    while confirmation != "1" or confirmation != "2" or confirmation == "0":#Asks user to confirm action of delete all
+    try:
+        confirmation = "0"
+        while confirmation != "1" or confirmation != "2" or confirmation == "0":#Asks user to confirm action of delete all
+            clear()
+            print("--------------Delete All Records------------------")
+            if checkIfNoRecords(userId) == 0:
+                print("\n**No records in database**\nPress any key to go back to Settings...")
+                getch()
+                return
 
-        print("--------------Delete Record------------------")
+            else:
+                print("\nAre you sure you want to delete all the records? (This action cannot be undone)\n")
+                print("[1] Yes")
+                print("[2] No, cancel and go back to Settings")
+                confirmation = input("\nSelection: ")
 
-        if checkIfNoRecords(userId) == 0:
-            print("\n**No records in database**")
-            break
+                if confirmation == "" or confirmation.isspace():
+                    print("\nSelection cannot be blank\nPress any key to try again...")
+                    getch()
+                    continue
 
-        else:
+                if confirmation == "1":
+                    delete_all_records(userId)
+                    print("Press any key to go back to Settings...")
+                    getch()
+                    return
 
-            print("\nAre you sure you want to delete all the records? (This action cannot be undone)\n")
-            print("[1] Yes")
-            print("[2] No, cancel and go back to Settings")
-            confirmation = input("\nSelection: ")
+                elif confirmation == "2":#If no then records will not be deleted
+                    print("\nNo records erased\nPress any key to go back to Settings...")
+                    getch()
+                    return
 
-            if confirmation == "":
-                print("\nSelection cannot be blank.")
-                break
+                else:#if invalid input
+                    print("\nInvalid Input! Please select a number between 1-2\nPress any key to try again...")
+                    getch()
+                    continue
 
-
-            if confirmation == "1":
-                delete_all_records(userId)
-                break
-
-
-
-            elif confirmation == "2":#If no then records will not be deleted
-                print("\nNo records erased")
-                break
-
-            else:#if invalid input
-                print("\nInvalid Input!")
-                break
-
-    print("Press any key to go back...")
-    getch()
-    clear()
-
+    except Exception:
+        print("\nInvalid Input")
+        print("Press any key to try again...")
+        getch()
+        clear()
 
 #Function allows user to change their email after confirming their password
 def changeEmail(userId):
     clear()
     while True:
         try:
-
-
-
             usrEmail = changeMasterEmail(userId)
 
             if usrEmail == '0':
-                break
-
-
+                return
 
             if usrEmail == False:
                 print("Press any key to go back...")
@@ -217,29 +231,25 @@ def changeEmail(userId):
                 break
 
             else:
-                print("Press any key to go back...")
-                getch()
+                if usrEmail == None:
+                    continue
+
+
                 clear()
                 break
 
-        except Exception:
-            print("\nInvalid Input. Error 6010")
+
+        except socket.gaierror:
+            print("\nConnection error. Please check your internet connection.")
             print("Press any key to try again...")
             getch()
             clear()
+            continue
 
 
-#Function used to change the user's master account password
-# def changePassword():
-#     clear()
-#     while True:
-#         try:
-#
-#             update_master_password(id)
-#
-#
-#         except Exception:
-#             print("\nInvalid Input.")
-#             print("Press any key to try again...")
-#             getch()
-#             clear()
+        except Exception:
+            print("\nInvalid Input")
+            print("Press any key to try again...")
+            getch()
+            clear()
+            continue
